Move transformSchema into separate file (#365)
